Revert "Triton norms dispatch refactor (#305)"
This reverts commit 78850fd73d40e8b283357540d2137a0f583ec99d.
