+++ tests/pytorch/triton_kernels/test_norm_common.py
@@ -1,11 +1,11 @@
-# Copyright (c) 2025-2026, Advanced Micro Devices, Inc. All rights reserved.
+# Copyright (c) 2025, Advanced Micro Devices, Inc. All rights reserved.
 # License for AMD contributions = MIT. See LICENSE for more information
 
 
 import torch
 from transformer_engine.pytorch import cpp_extensions as tex
 from transformer_engine.pytorch.triton_kernels.common import torch_dtype_to_te_dtype
-from transformer_engine.pytorch.triton_kernels.utils import get_num_sms
+from transformer_engine.pytorch.triton_kernels.norm_common import get_num_sms
 
 def test_sm_margin():
     num_sms = get_num_sms()
