+++ transformer_engine/pytorch/module/_common.py
@@ -1,5 +1,5 @@
 # This file was modified for portability to AMDGPU
-# Copyright (c) 2024-2026, Advanced Micro Devices, Inc. All rights reserved.
+# Copyright (c) 2024-2025, Advanced Micro Devices, Inc. All rights reserved.
 # Copyright (c) 2022-2025, NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 #
 # See LICENSE for license information.
@@ -20,12 +20,8 @@ from ..utils import get_default_init_method
 from ..export import is_in_onnx_export_mode
 
 if IS_HIP_EXTENSION:
-    from ..triton_kernels.norms_common import (
-        te_layernorm_fwd_triton,
-        te_layernorm_bwd_triton,
-        te_rmsnorm_fwd_triton,
-        te_rmsnorm_bwd_triton
-    )
+    from ..triton_kernels.layernorm import te_layernorm_fwd_triton, te_layernorm_bwd_triton
+    from ..triton_kernels.rmsnorm import te_rmsnorm_bwd_triton, te_rmsnorm_fwd_triton
 
 def _get_normalization_func(normalization: str, forward: bool):
     use_rmsnorm_triton = bool( int(os.environ.get('NVTE_USE_RMSNORM_TRITON', '0')) ) and IS_HIP_EXTENSION
