+++ transformer_engine/pytorch/module/layernorm_linear.py
@@ -79,7 +79,8 @@ from ..cpp_extensions import (
 )
 
 if IS_HIP_EXTENSION:
-    from ..triton_kernels.norms_common import te_layernorm_bwd_triton, te_rmsnorm_bwd_triton
+    from ..triton_kernels.layernorm import te_layernorm_bwd_triton
+    from ..triton_kernels.rmsnorm import te_rmsnorm_bwd_triton
 
 
 __all__ = ["LayerNormLinear"]
