+++ transformer_engine/pytorch/triton_kernels/norm_common.py
@@ -1,4 +1,4 @@
-# Copyright (c) 2025-2026, Advanced Micro Devices, Inc. All rights reserved.
+# Copyright (c) 2025, Advanced Micro Devices, Inc. All rights reserved.
 # License for AMD contributions = MIT. See LICENSE for more information
 
 import os
