+++ transformer_engine/pytorch/triton_kernels/rmsnorm.py
@@ -1,31 +1,44 @@
-# Copyright (c) 2025-2026, Advanced Micro Devices, Inc. All rights reserved.
+# Copyright (c) 2025, Advanced Micro Devices, Inc. All rights reserved.
 # License for AMD contributions = MIT. See LICENSE for more information
 
 import torch
 import triton
 import triton.language as tl
 from itertools import product
+from .norm_common import num_programs, block_size, use_blocked, make_ln_out
+from transformer_engine.pytorch.tensor.float8_tensor import Float8Quantizer, Float8CurrentScalingQuantizer
+from transformer_engine.pytorch.tensor.mxfp8_tensor import MXFP8Quantizer
+from transformer_engine.pytorch.triton_kernels.common import (
+    te_dtype_to_torch_dtype,
+    te_dtype_to_triton_dtype,
+)
+from .common import get_fp8_max
+from ..tensor.quantized_tensor import Quantizer
+import transformer_engine_torch as tex
+
+def dg_tmp_rows(x, sm_margin=None):
+    return x.shape[0] if use_blocked(x) else num_programs(x, sm_margin)
+
 
 def get_autotune_config():
     return [triton.Config({'waves_per_eu': we}, num_warps=nw) for (we, nw) in product([0, 1, 2, 4], [4, 8, 16])]
 
 
-# TODO(micky774) Implement fused MXFP8 quantization within the kernel
 @triton.jit
 def _rmsnorm_fwd_triton_impl(
-    input_ptr,
     output_ptr,
-    g_ptr,
-    rsigma_ptr,
+    input_ptr,
+    g_ptr, rsigma_ptr,
     input_row_stride,
     output_row_stride,
     n_rows, n_cols,
     epsilon,
+    amax_ptr,
     q_amax_ptr,
     q_scale_ptr,
     scale_inv_ptr,
     out_transpose_ptr,
-    out_transpose_stride,
+    transpose_row_stride,
     ZERO_CENTERED_GAMMA: tl.constexpr,
     BLOCK_SIZE: tl.constexpr,
     USE_BLOCKED: tl.constexpr,
@@ -109,7 +122,7 @@ def _rmsnorm_fwd_triton_impl(
                     rms_norm = rms_norm * scale
                     rms_norm = tl.clamp(rms_norm, -FP8_MAX, FP8_MAX)
                     if MAKE_TRANSPOSE:
-                        output_t_ptrs = out_transpose_ptr + cols * out_transpose_stride + row_idx
+                        output_t_ptrs = out_transpose_ptr + cols * transpose_row_stride + row_idx
                         tl.store(output_t_ptrs, rms_norm.to(output_type))
                 tl.store(output_ptrs, rms_norm.to(output_type))
 
@@ -134,7 +147,7 @@ def _rmsnorm_fwd_triton_impl(
                 rms_norm = rms_norm * scale
                 rms_norm = tl.clamp(rms_norm, -FP8_MAX, FP8_MAX)
                 if MAKE_TRANSPOSE:
-                    output_t_ptrs = out_transpose_ptr + cols * out_transpose_stride  + row_idx
+                    output_t_ptrs = out_transpose_ptr + cols * transpose_row_stride  + row_idx
                     tl.store(output_t_ptrs, rms_norm.to(output_type), mask=mask)
             tl.store(output_ptrs, rms_norm.to(output_type), mask=mask)
 
@@ -167,10 +180,11 @@ def _rmsnorm_fwd_triton_impl(
                 rms_norm = rms_norm * scale
                 rms_norm = tl.clamp(rms_norm, -FP8_MAX, FP8_MAX)
                 if MAKE_TRANSPOSE:
-                    output_t_ptrs = out_transpose_ptr + col_offsets * out_transpose_stride + row_idx
+                    output_t_ptrs = out_transpose_ptr + col_offsets * transpose_row_stride + row_idx
                     tl.store(output_t_ptrs, rms_norm.to(output_type), mask=mask)
             tl.store(output_ptrs, rms_norm.to(output_type), mask=mask)
     if IS_FP8:
+        tl.store(amax_ptr + row_start, amax)
         tl.atomic_max(q_amax_ptr, amax, sem="relaxed")
         if row_start == 0:
             scale = tl.load(q_scale_ptr)
@@ -349,3 +363,165 @@ def _rmsnorm_bwd_dg_reduce_triton(dg_in_ptr, dg_out_ptr, dg_in_stride, n_rows, n
     sum_dg = tl.sum(acc, axis=0)
     tl.store(dg_out_ptr + cols, sum_dg.to(dg_out_ptr.type.element_ty), mask=cols < n_cols)
 
+# triton drop-in replacement for transformer_engine::pytorch::rmsnorm_bwd
+def te_rmsnorm_bwd_triton(dz, x, rsigma, gamma, sm_margin, zero_centered_gamma):
+    # may take non-contiguous inputs
+    dz_ = dz.contiguous()
+    x_ = x.contiguous()
+    rsigma_ = rsigma.contiguous()
+    gamma_ = gamma.contiguous()
+
+    dx = torch.empty_like(x_)
+    dgamma = torch.empty_like(gamma_)
+
+    M, N = x_.shape
+    blk_size = block_size(x_)
+    USE_BLOCKED = use_blocked(x_)
+    NUM_PRGMS = num_programs(x_, sm_margin)
+    need_reduction = N > 1
+    dg_tmp = torch.empty(dg_tmp_rows(x_, sm_margin), N, device=x.device, dtype=torch.float32, requires_grad=False) if need_reduction else None
+
+    grid_bwd = lambda meta: (NUM_PRGMS, )
+    input_aligned_16 = (x_.data_ptr() % 16 == 0) and (x_.stride(-1) % 16 == 0)
+    grad_output_aligned_16 = (dz_.data_ptr() % 16 == 0) and (dz_.stride(-1) % 16 == 0)
+    dx_aligned_16 = (dx.data_ptr() % 16 == 0) and (dx.stride(-1) % 16 == 0)
+    dg_target = dg_tmp if need_reduction else dgamma
+    dg_aligned_16 = (dg_target.data_ptr() % 16 == 0) and (dg_target.stride(-1) % 16 == 0)
+    _rmsnorm_bwd_triton[grid_bwd](
+        dz_,
+        x_,
+        gamma_,
+        rsigma_,
+        dx,
+        dg_target,
+        x_.stride(0),
+        dz_.stride(0),
+        M,
+        N,
+        zero_centered_gamma,
+        blk_size,
+        USE_BLOCKED,
+        NUM_PRGMS,
+        input_aligned_16,
+        grad_output_aligned_16,
+        dx_aligned_16,
+        dg_aligned_16,
+        num_warps=8,
+    )
+
+    if need_reduction:
+        grid_reduce = lambda meta: [triton.cdiv(N, meta['BLOCK_SIZE_N'])]
+        _rmsnorm_bwd_dg_reduce_triton[grid_reduce](dg_tmp, dgamma, dg_tmp.stride(0), dg_tmp.shape[0], dg_tmp.shape[1],
+                                                   BLOCK_SIZE_M=128, BLOCK_SIZE_N=64)
+
+    return dx, dgamma
+
+# triton drop-in replacement for transformer_engine::pytorch::rmsnorm_fwd
+def te_rmsnorm_fwd_triton(
+    input: torch.Tensor,
+    weight: torch.Tensor,
+    eps: float,
+    ln_out: torch.Tensor,
+    quantizer: Quantizer,
+    otype: tex.DType,
+    sm_margin: int,
+    zero_centered_gamma: bool,
+    autotune: bool = True,
+):
+    if eps < 0:
+        raise ValueError(f"`eps` must be non-negative, but a value of {eps} was passed")
+    if len(input.shape) != 2:
+        raise ValueError(
+            f"The input must be a 2-dimensional matrix, but an input with {input.ndim} was passed.")
+
+    device = input.device
+    N, H = input.shape
+    if weight.shape[0] != H:
+        raise ValueError(
+            f"The shape of `weight` must be feature-aligned, "
+            f"but {weight.shape[0]=} while {input.shape[1]=}"
+        )
+    IS_FP8 = isinstance(quantizer, Float8Quantizer)
+    IS_FP8_CURRENT_SCALING = isinstance(quantizer, Float8CurrentScalingQuantizer)
+    IS_MXFP8 = isinstance(quantizer, MXFP8Quantizer)
+    BLOCK_SIZE = block_size(input)
+    USE_BLOCKED = use_blocked(input)
+    NUM_PRGMS = num_programs(input, sm_margin)
+    MAKE_TRANSPOSE = False
+
+    rsigma = torch.empty((N,), dtype=torch.float32, device=device)
+    torch_out_dtype = (
+        otype if isinstance(otype, torch.dtype)
+        else te_dtype_to_torch_dtype(otype)
+    )
+    out = make_ln_out(
+        ln_out,
+        quantizer=quantizer,
+        input_shape=input.shape,
+        out_dtype=torch_out_dtype
+    )
+    if IS_FP8:
+        MAKE_TRANSPOSE = quantizer.columnwise_usage
+        amax = torch.empty((NUM_PRGMS,), dtype=torch.float32, device=device)
+        tl_dtype = te_dtype_to_triton_dtype(quantizer.dtype)
+        scale_inv_ptr = out._scale_inv
+        q_scale = quantizer.scale
+        q_amax = quantizer.amax
+        out_ptr = triton.reinterpret(out._data, tl_dtype)
+        FP8_MAX = get_fp8_max(quantizer.dtype)
+        if MAKE_TRANSPOSE:
+            if out._transpose_invalid:
+                out._transpose = torch.empty((out._data.shape[1], out._data.shape[0]), dtype=out._data.dtype, device=device)
+                out._transpose_invalid = False
+            out_transpose_ptr = triton.reinterpret(out._transpose, tl_dtype)
+            out_transpose_stride = out._transpose.stride(0)
+        else:
+            out_transpose_ptr = None
+            out_transpose_stride = None
+    else:
+        amax = None
+        tl_dtype = None
+        scale_inv_ptr = None
+        q_scale = None
+        q_amax = None
+        out_ptr = out
+        out_transpose_ptr = None
+        out_transpose_stride = None
+        FP8_MAX = None
+
+    grid_fwd = lambda meta: (NUM_PRGMS, )
+    # TODO(micky774) Implement fused MXFP8 quantization within the kernel
+    kernel = _rmsnorm_fwd_triton if autotune else _rmsnorm_fwd_triton_impl
+    input_aligned_16 = (input.data_ptr() % 16 == 0) and (input.stride(-1) % 16 == 0)
+    out_alignment_tensor = out._data if hasattr(out, "_data") else out
+    output_aligned_16 = (out_alignment_tensor.data_ptr() % 16 == 0) and (
+        out_alignment_tensor.stride(-1) % 16 == 0
+    )
+    kernel[grid_fwd](
+        out_ptr,
+        input,
+        weight,
+        rsigma,
+        input.stride(0),
+        out_ptr.stride(0),
+        N, H, eps,
+        amax,
+        q_amax,
+        q_scale,
+        scale_inv_ptr,
+        out_transpose_ptr,
+        out_transpose_stride,
+        zero_centered_gamma,
+        BLOCK_SIZE,
+        USE_BLOCKED,
+        NUM_PRGMS,
+        IS_FP8,
+        FP8_MAX,
+        MAKE_TRANSPOSE,
+        input_aligned_16,
+        output_aligned_16,
+    )
+    if IS_MXFP8 or IS_FP8_CURRENT_SCALING:
+        out = quantizer.quantize(out, out=ln_out)
+
+    return out, None, rsigma
